requirements.txt
Django==2.2.17
django-otp
django-simple-captcha
qrcode
